% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{crop_model_methods}
\alias{crop_model_methods}
\title{Available statistical crop-yield methods}
\usage{
crop_model_methods()
}
\value{
A character vector of method identifiers accepted by
  [crop_model_config()] and [fit_crop_model()].
}
\description{
Available statistical crop-yield methods
}
\examples{
crop_model_methods()
}
