% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{yieldlag-package}
\alias{yieldlag}
\alias{yieldlag-package}
\title{YieldLag: seasonal climate responses and regional crop yields}
\description{
The package supplies a deterministic, base-R modelling engine and a
high-level workflow that keeps tuning, later evaluation, final fitting, and
artifact export separate. See [crop_model_config()], [read_crop_data()],
[compare_crop_models()], and [write_crop_results()] for the main workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MR-Eini/yieldlag}
  \item Report bugs at \url{https://github.com/MR-Eini/yieldlag/issues}
}

}
\author{
\strong{Maintainer}: Mohammad Reza Eini \email{mohammad_eini@sggw.edu.pl}

Authors:
\itemize{
  \item Mohammad Reza Eini \email{mohammad_eini@sggw.edu.pl}
}

}
\keyword{internal}
